// src/features/booking/dto/booking.dto.ts
export interface BookingDto {
  bookingId: number;
  customerId: number;
  location: string;
  date: string;
  createdAt: string;
  status: string;
  amount: number;
  skintherapistId: number;
  serviceName: string;
  serviceId: number;
  note: string;
  skintherapistName: string;
  customer: string;
  updateAt: string;
}